Fix: 코드리뷰 반영

+++ src/main/java/com/team3/deokhugam/batch/dto/PopularBookRawData.java
@@ -1,9 +1,10 @@
 package com.team3.deokhugam.batch.dto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record PopularBookRawData (
     UUID bookId,
     int reviewCount,
-    double ratingAvg
+    BigDecimal ratingAvg
 ){}
